Support self-GET in comm=ofi.
There was a sanity check against GET-from-self in the ofi comm layer,
but it turns out that we actually do that under certain circumstances.
An example is the test arrays/bradc/slices/fftSlice, which is run on
just a single locale.  In the future it would be nice to catch all the
cases of self-GET upstream of the comm layers, but for now since this is
supported by comm=gasnet and comm=ugni I'm enabling it in comm=ofi as
well.

While here I also fixed a mistyped function argument in the "sockets"
out-of-band support.  This changed a few weeks ago in the other (PMI)
OOB layer but I didn't catch the sockets one at the time.

+++ runtime/src/comm/ofi/comm-ofi-oob-sockets.c
@@ -85,7 +85,7 @@ void chpl_comm_ofi_oob_barrier(void) {
 }
 
 
-void chpl_comm_ofi_oob_allgather(void* in, void* out, int len) {
+void chpl_comm_ofi_oob_allgather(void* in, void* out, size_t len) {
   if (chpl_numNodes == 1) {
     chpl_memcpy(out, in, len);
   } else {
